+++ fabric/src/main/kotlin/dev/nyon/magnetic/config/Identifier.kt
@@ -3,5 +3,9 @@ package dev.nyon.magnetic.config
 import kotlinx.serialization.Serializable
 import net.minecraft.resources.ResourceLocation
 
-@Serializable(with = ResourceLocationSerializer::class)
-data class Identifier(val original: ResourceLocation, val isTag: Boolean)
+@Serializable(with = IdentifierSerializer::class)
+data class Identifier(val original: ResourceLocation, val isTag: Boolean) {
+    override fun toString(): String {
+        return "${if (isTag) "#" else ""}$original"
+    }
+}