+++ changelog.md
@@ -1,3 +1,8 @@
 # Both
+
 - add option to require a player to have a permission ('magnetic.ability.use') to use the ability
-  - on Fabric the fabric-permissions-api has to be installed for this to work
+    - on Fabric the fabric-permissions-api has to be installed for this to work
+- add option to ignore drops from entities when killed
+    - you can use identifiers and tags to define these entities
+    - the option is called 'ignoreKilledEntities' and can be configured via the config file or in-game in the config
+      screen